Moved run reducers to admin actions file
