+++ src/reducers/panelMeetingAdmin/index.js
@@ -1,7 +1,23 @@
-import { createPanelMeetingHasErrored, createPanelMeetingIsLoading, createPanelMeetingSuccess } from './createPanelMeeting';
+import {
+  createPanelMeetingHasErrored,
+  createPanelMeetingIsLoading,
+  createPanelMeetingSuccess,
+  runOfficialAddendumErrored,
+  runOfficialAddendumSuccess,
+  runOfficialPreliminaryErrored,
+  runOfficialPreliminarySuccess,
+  runPostPanelProcessingErrored,
+  runPostPanelProcessingSuccess,
+} from './panelMeetingAdmin';
 
 export default {
   createPanelMeetingHasErrored,
   createPanelMeetingIsLoading,
   createPanelMeetingSuccess,
+  runOfficialAddendumErrored,
+  runOfficialAddendumSuccess,
+  runOfficialPreliminaryErrored,
+  runOfficialPreliminarySuccess,
+  runPostPanelProcessingErrored,
+  runPostPanelProcessingSuccess,
 };
