+++ src/reducers/panelMeetings/index.js
@@ -9,12 +9,6 @@ import {
   panelMeetingsFiltersFetchDataErrored,
   panelMeetingsFiltersFetchDataLoading,
   panelMeetingsSelections,
-  runOfficialAddendumErrored,
-  runOfficialAddendumSuccess,
-  runOfficialPreliminaryErrored,
-  runOfficialPreliminarySuccess,
-  runPostPanelProcessingErrored,
-  runPostPanelProcessingSuccess,
 } from './panelMeetings';
 
 export default {
@@ -28,10 +22,4 @@ export default {
   panelMeetingsFiltersFetchDataErrored,
   panelMeetingsFiltersFetchDataLoading,
   panelMeetingsSelections,
-  runOfficialAddendumErrored,
-  runOfficialAddendumSuccess,
-  runOfficialPreliminaryErrored,
-  runOfficialPreliminarySuccess,
-  runPostPanelProcessingErrored,
-  runPostPanelProcessingSuccess,
 };
