+++ src/reducers/panelMeetings/panelMeetings.js
@@ -89,54 +89,3 @@ export function panelMeeting(state = {}, action) {
       return state;
   }
 }
-
-// ======================== Panel Meeting Actions ========================
-
-export function runOfficialPreliminaryErrored(state = false, action) {
-  switch (action.type) {
-    case 'RUN_OFFICIAL_PRELIMINARY_ERRORED':
-      return action.hasErrored;
-    default:
-      return state;
-  }
-}
-export function runOfficialPreliminarySuccess(state = {}, action) {
-  switch (action.type) {
-    case 'RUN_OFFICIAL_PRELIMINARY_SUCCESS':
-      return action.results;
-    default:
-      return state;
-  }
-}
-export function runOfficialAddendumErrored(state = false, action) {
-  switch (action.type) {
-    case 'RUN_OFFICIAL_ADDENDUM_ERRORED':
-      return action.hasErrored;
-    default:
-      return state;
-  }
-}
-export function runOfficialAddendumSuccess(state = {}, action) {
-  switch (action.type) {
-    case 'RUN_OFFICIAL_ADDENDUM_SUCCESS':
-      return action.results;
-    default:
-      return state;
-  }
-}
-export function runPostPanelProcessingErrored(state = false, action) {
-  switch (action.type) {
-    case 'RUN_POST_PANEL_PROCESSING_ERRORED':
-      return action.hasErrored;
-    default:
-      return state;
-  }
-}
-export function runPostPanelProcessingSuccess(state = {}, action) {
-  switch (action.type) {
-    case 'RUN_POST_PANEL_PROCESSING_SUCCESS':
-      return action.results;
-    default:
-      return state;
-  }
-}
